fix(shibboleth.idp): workaround permissions issue, add /bin/bash symlink on FreeBSD
The file.recurse state does not copy file system permissions from the
master to the minion (which in any case get lost by git), so work around
this by explicitly calling shell scripts with `/bin/sh`.

The various scripts included in the IdP distfile assume that GNU bash is
installed on the root file system.  This is not the case on operating
systems such as FreeBSD, which do not install bash by default and which
install bash under folders like `/opt` or `/usr/local`.  Work around
this by creating a symlink for `/bin/bash` that points to the
appropriate executable.