refactor(shibboleth.idp): clean up if statements/expressions
In order to improve readability, this restructures inline if expressions
vertically:

* Move values to a new line and indent them an extra level.

* Move inline if expressions to a new line under the value and indented
  two extra levels if they flow past the 80th column.

* Move clauses within if expressions to their own lines and indent them
  to match their parent if expression if together they flow past the
  80th column.

* Move boolean operators to the beginning of each clause.

This also adjust if statements to match.

In a few cases, this change simplifies the inline if expression or
removes it entirely.  Likewise, in several cases this change re-orders
conditionals in order to more easily support string or integer values,
as opposed to adding additional if statements/expressions.