fix(shibboleth.idp): switch to `onchanges` requisites
SaltStack has deprecated the `cmd.wait` and similar states in favor of
the more general `onchanges` requisite.

Likewise, this changes the behavior of the `cmd: shibidp_keymat` state
to only watch changes to the IdP keying material.  PKCS#12 container
regeneration due to keystore password changes must be signaled by
another state as described in the (revised as part of this change)
comments.